pyproject.toml
[project]
name = "orchestrator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.4",
    "aiohttp>=3.12.15",
    "colorlog>=6.9.0",
    "descope>=1.7.9",
    "fastapi>=0.116.1",
    "google-adk>=1.13.0",
    "litellm>=1.76.2",
    "psycopg[binary]>=3.2.9",
    "redis>=6.4.0",
]
